srv connect

+++ src/lin/srv/cfg.go
@@ -11,7 +11,7 @@ import (
 type ServerCfg struct {
 	Cluster string `yaml:"cluster"`
 	BindAddr string `yaml:"bind_addr"`
-	SrvID int `yaml:"srv_id"`
+	SrvID int64 `yaml:"srv_id"`
 	AliasName string `yaml:"alias_name"`
 	HttpAddr string `yaml:"http_addr"`
 	MapCluster map[int]string `yaml:"map_cluster"`
